+++ R/accessors.R
@@ -46,6 +46,7 @@ setMethod("errlogfile", "GRANRepository", function(repo) {
 #' @aliases errlogfile,RepoBuildParam-method
 setMethod("errlogfile", "RepoBuildParam", function(repo) repo@errlog)
 
+
 #' Retrieve the path to a GRAN (sub) repository
 #' @rdname location-methods
 #' @param repo A GRANRepository object.
@@ -111,7 +112,8 @@ setMethod("repobase", "GRANRepository", function(repo) {
 
 
 #' staging
-#' Return the staging directory or the staging_logs to be used when building the repository. If the directory does not exist it will be created.
+#' Return the staging directory or the staging_logs to be used when building
+#' the repository. If the directory does not exist it will be created.
 #'
 #' @rdname staging-methods
 #' @param repo a GRANRepository object
@@ -184,10 +186,26 @@ setGeneric("destination", function(repo) standardGeneric("destination"))
 #' @rdname destination-methods
 #' @aliases destination,GRANRepository-method
 #' @export
-setMethod("destination","GRANRepository",
+setMethod("destination", "GRANRepository",
           function(repo) file.path(normalizePath2(param(repo)@dest_base),
                                    param(repo)@repo_name,  "src", "contrib"))
 
+#' archivedir
+#' Return the full path to the archive directory for the final repository
+#' deployment.
+#'
+#' @rdname archivedir-methods
+#' @param repo a GRANRepository object
+#' @return The full path to the archive directory where the archived packages
+#' will be deployed to
+#' @docType methods
+#' @export
+setGeneric("archivedir", function(repo) standardGeneric("archivedir"))
+#' @rdname archivedir-methods
+#' @aliases archivedir,GRANRepository-method
+#' @export
+setMethod("archivedir", "GRANRepository",
+          function(repo) file.path(destination(repo), "Archive"))
 
 #' dest_base
 #' Return the full path to the contrib directory for the final repository