+++ R/addRevDeps.R
@@ -15,7 +15,7 @@ addRevDeps = function(repo)
         repo_results(repo)$buildReason[is.na(repo_results(repo)$buildReason)] = ""
         repo_results(repo)$buildReason[repo_results(repo)$building] = "vbump"
     }
-    
+
     logfun(repo)("NA", paste("Checking for reverse dependencies to packages",
                              "with version bumps."))
     manifest = repo_results(repo)
@@ -28,7 +28,7 @@ addRevDeps = function(repo)
                                  "reverse dependency check."))
         return(repo)
     }
-    
+
     pkgs = manifest$name[getBuilding(repo)]
     revdeps = sapply(pkgs, dependsOnPkgs, dependencies = "all",
         lib.loc = temp_lib(repo), simplify=FALSE)
@@ -62,31 +62,14 @@ addRevDeps = function(repo)
 
     rdepBuildPkgs = manifest$buildReason != "vbump" & !(manifest$name %in% suspended_pkgs(repo)) & grepl(paste0("(", paste(pkgs, collapse="|"), ")"), manifest$revDepOf)
     manifest[rdepBuildPkgs, "buildReason"] = "is rdep"
-    manifest[rdepBuildPkgs, "building"] = TRUE 
-    
+    manifest[rdepBuildPkgs, "building"] = TRUE
+
     logfun(repo)("NA", paste0("Detected ", sum(rdepBuildPkgs),
                               " packages that are reverse dependencies of",
                               "packages with version bumps:\n\t",
                               paste(manifest$name[rdepBuildPkgs],
                                     collapse=" , ")))
 
-## I don't think we need this because this is happening before the main
-    ## buildBranchesInRepo call....
-    ##    if(sum(rdepBuildPkgs) >0) {
-    ##        logfun(repo)("NA",
-    ##                     "Building reverse dependencies in temporary repository.")
-    ##        tmprepo = repo
-    ##       repo_results(tmprepo) = manifest[rdepBuildPkgs,]
-    ##       manifest_df(tmprepo) = manifest_df(tmprepo)[rdepBuildPkgs,]
-    ##       versions_df(tmprepo) = versions_df(tmprepo)[rdepBuildPkgs,]
-    ##       repo_results(tmprepo)$building = TRUE
-    ##       repo_results(tmprepo)$status="ok"
-    ##       tmprepo = buildBranchesInRepo(repo = tmprepo, temp = TRUE,
-                                        # incremental = FALSE)
-    ##           incremental = TRUE)
-    ##       manifest[rdepBuildPkgs, ] = repo_results(tmprepo)
-    ##   }
     repo_results(repo) = manifest
     repo
 }
-