+++ R/getRVersion.R
@@ -1,5 +1,0 @@
-getRVersion <- function() {
-  paste0(R.Version()[["major"]],
-         ".",
-         R.Version()[["minor"]])
-}