+++ R/pkgHTML.R
@@ -4,11 +4,11 @@
 #' @importFrom tools file_path_sans_ext
 #' @importFrom jsonlite toJSON
 #' @param repo A gRAN repo object
-#' @param suffix Append a suffix to the HTML splash page
+#' @param splashname Filename for the HTML splash page
 #' @param theme CSS theme. bootstrap, foundation, semanticui or jqueryui
 #' @return None
 #' @export
-pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
+pkgHTML <- function(repo, splashname = "index.html", theme = "bootstrap") {
   logfun(repo)("NA", paste0("Creating HTML splash pages for ",
                             sum(repo_results(repo)$building), " packages"),
                             type = "full")
@@ -24,6 +24,9 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
       return(repo)
   }
   bres <- subset(bres, !(grepl("^GRAN", bres$name)))
+
+  # Get manifest_df for SCM info
+  scm_df <- manifest_df(repo)
 
   # Copy the stylesheet
   stylesheet <- paste0("<link rel=\"stylesheet\" type=\"text/css\"
@@ -46,16 +49,11 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
         descr_df <- generateDescInfo(file.path(check_dir, pkg_name))
 
         # Create JSON of the DESCRIPTION file
-        reponame <- paste0("GRAN", repo_name(repo))
-        descr_df$id <- encode_string(paste0(reponame, descr_df$Package))
-        descr_df$GranRepo <- reponame
-        desc_json <- toJSON(descr_df, pretty = TRUE)
-        json_outfile <- file.path(docdir, paste0(descr_df$Package, "-desc.json"))
-        write(desc_json, json_outfile)
+        createJSON(repo, pkg_name, descr_df, scm_df, docdir)
 
         # Exclude these fields from the splash page
         descr_df <- descr_df[ , !(names(descr_df) %in%
-                                  c("Authors@R", "Collate", "id", "GranRepo"))]
+                                  c("Authors@R", "Collate"))]
         if("URL" %in% colnames(descr_df)) {
           descr_df$URL <- createURL(descr_df$URL)
         }
@@ -84,10 +82,10 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
       if (!(is.data.frame(revdeps) && ncol(revdeps)==0)) {
         revdeps_header <- "<br/><h4>Reverse Dependencies</h4><hr>"
         revdeps_html <- htmlTable(t(revdeps),
-                          css.cell = ("padding-left: 0.5em; padding-right: 0.5em"),
-                          css.class = "table table-striped table-hover",
-                          css.table = "margin-left:10px;margin-right:10px;",
-                          align="l", rnames = names(revdeps))
+                        css.cell = ("padding-left: 0.5em; padding-right: 0.5em"),
+                        css.class = "table table-striped table-hover",
+                        css.table = "margin-left:10px;margin-right:10px;",
+                        align="l", rnames = names(revdeps))
         logfun(repo)(pkg_name, "Created revdep info", type = "full")
       } else {
         revdeps_header <- ""
@@ -105,7 +103,7 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
       authors <- emailTag(as.character(descr_df$Author))
       title <- as.character(descr_df$Title)
       intro <- paste0("<h2>", pkg_name, ": ",title, "</h2><hr> ",
-              #"<img src=\"", pkg_name, ".png\" height=\"160\" align=\"right\">",
+              "<img src=\"", pkg_name, ".png\" height=\"160\" align=\"right\">",
                       "<strong><p>", description, "</p></strong> ",
                       "<p>GRAN Release: ",
                       "<a href=\"../../src/contrib/buildreport.html\">",
@@ -120,6 +118,9 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
       # Installation instructions
       installn <- paste0("<br/><h4>Installation</h4><hr>",
                          "<p>To install this package, start R and enter:</p>",
+                         #"<pre><code>install.packages(\"", pkg_name,
+                         #"\", repos = \"", param(repo)@dest_url, "/",
+                         #repo_name(repo), "\")</code></pre>")
                          "<pre><code><p>source(\"", param(repo)@dest_url,
                          "/getGRAN-", repo_name(repo),
                          ".R\")</p>", "<p>library(GRAN", repo_name(repo),")</p>",
@@ -213,7 +214,7 @@ pkgHTML <- function(repo, suffix = "-index.html", theme = "bootstrap") {
 
       # Create the HTML spash page
       logfun(repo)(pkg_name, "Writing final pkg HTML info", type = "full")
-      write(final_html, file = file.path(docdir, paste0(pkg_name, suffix)))
+      write(final_html, file = file.path(docdir, splashname))
     }
   }
 
@@ -344,8 +345,7 @@ extPkgURL <- function(desc_field, as_string = TRUE) {
 #' @param relation What type of reverse dependency?
 #' @return String of reverse dependencies
 relatedPkgs <- function(pkg_name, relation = "LinkingTo") {
-  paste(dependsOnPkgs(pkg_name, dependencies = relation,
-                             recursive = FALSE),
+  paste(dependsOnPkgs(pkg_name, dependencies = relation, recursive = FALSE),
         collapse = ', ')
 }
 
@@ -356,8 +356,8 @@ relatedPkgs <- function(pkg_name, relation = "LinkingTo") {
 #' @param separator The delimiter
 #' @return Processed list
 string2list <- function(string, separator = ",") {
-  as.list(strsplit(stri_replace_all_charclass(string,
-                    "\\p{WHITE_SPACE}", ""), separator)[[1]])
+  as.list(strsplit(stri_replace_all_charclass(string, "\\p{WHITE_SPACE}", ""),
+          separator)[[1]])
 }
 
 #' Creates a href tag
@@ -436,4 +436,41 @@ buildBadge <- function(status, pkg_name) {
   }
   paste("<a href=\"", log, "\"><span class=\"label",
         label, "\">", status, "</span></a>")
+}
+
+#' Create JSON representation of package information
+#' @author Dinakar Kulkarni <kulkard2@gene.com>
+#' @importFrom jsonlite toJSON
+#' @param repo A GRAN repo object
+#' @param pkg_name Name of the GRAN package
+#' @param descr_df data.frame representation of DESCRIPTION file
+#' @param scm_df data.frame representation of GRAN manifest object
+#' @param docdir Directory where the JSON doc will be written
+#' @param suffix Suffix for the JSON file
+#' @return None. Write JSON file to disk
+#' @seealso \code{\link{manifest_df}} for generating scm_df and
+#'    \code{\link{generateDescInfo}} for generating descr_df.
+createJSON <- function(repo, pkg_name, descr_df, scm_df, docdir,
+                       suffix = paste0("_", descr_df$Version, "-desc.json")) {
+  reponame <- paste0("GRAN", repo_name(repo))
+  descr_df$id <- encode_string(paste0(reponame, descr_df$Package))
+  descr_df$GranRepo <- reponame
+
+  # Get SCM info:
+  scm_info <- scm_df[scm_df$name == pkg_name, ]
+  descr_df$scm_url <- scm_info$url
+  descr_df$scm_type <- scm_info$type
+  descr_df$scm_branch <- scm_info$branch
+  descr_df$scm_subdir <- scm_info$subdir
+
+  # Get documentation URL locations:
+  doc_url <- paste0(repo_url(repo), "/PkgDocumentation/", pkg_name)
+  descr_df$pkgdocs_url <- doc_url
+
+  # Convert to JSON
+  desc_json <- toJSON(descr_df, pretty = TRUE)
+  json_outfile <- file.path(docdir, paste0(pkg_name, suffix))
+
+  # Write JSON
+  write(desc_json, json_outfile)
 }